Version 5

+++ app/guildconfig.php
@@ -27,8 +27,6 @@ include("include/guild_request.php");
 if(!IsSet($accept))	{
 }
 else	{	
-
-	include("include/connect.php");
 	
 	// retrieve necessary guild info
 	$guild_info_query = mysql_db_query($dbnam, "SELECT * FROM guild WHERE owner='$userid'");
@@ -39,12 +37,15 @@ else	{
 	// get their empire name
 	$applicant_ename_query = mysql_db_query($dbnam, "SELECT ename FROM user WHERE userid='$auserid'");
 		$applicant_ename = mysql_fetch_array($applicant_ename_query);
+	// get their current guild
+	$current_guild_query = mysql_db_query($dbnam, "SELECT guild FROM user WHERE userid='$auserid'");
+		$current_guild = mysql_fetch_array($current_guild_query);
 
-	if($guild_members > 29)	{
+	if($guild_members > 14)	{
 		echo "<div align=center><font class=yellow>Your guild is full! You cannot allow any more members in!</font></div>";
 		die();
 	}
-	else	{
+	elseif($current_guild[guild] == None)	{
 
 		mysql_query("UPDATE user SET guild='$guild_info[gname]' WHERE userid='$auserid'");
 		mysql_query("INSERT INTO empnews (date, news, yourid) VALUES	('$clock', '<font class=blue>You were accepted into $guild_info[gname].</font>', '$auserid') ");
@@ -59,25 +60,34 @@ else	{
 		echo "<div align=center><font class=yellow><b>$applicant_ename[ename] has been accepted into the guild!</b></font></div>";
 		die();
 	}
+	else	{
+		echo "<div align=center><font class=yellow>That person is curently in a guild!</font></div>";
+		die();
+	}
 }
 
 if(!IsSet($reject))	{
 }
 else	{	
-
-	include("include/connect.php");
+	
 	// retrieve necessary guild info
 	$guild_info_query = mysql_db_query($dbnam, "SELECT * FROM guild WHERE owner='$userid'");
 		$guild_info = mysql_fetch_array($guild_info_query);
 	// get their empire name
 	$applicant_ename_query = mysql_db_query($dbnam, "SELECT ename FROM user WHERE userid='$auserid'");
 		$applicant_ename = mysql_fetch_array($applicant_ename_query);
+	// get their current guild
+	$current_guild_query = mysql_db_query($dbnam, "SELECT guild FROM user WHERE userid='$auserid'");
+		$current_guild = mysql_fetch_array($current_guild_query);
 	
-	mysql_query("INSERT INTO empnews (date, news, yourid) VALUES	('$clock', '<font class=blue>You were rejected from $guild_info[gname].</font>', '$auserid') ");
-	mysql_query("DELETE FROM guildrequests WHERE applicant='$auserid' AND gl_userid='$userid'");
+	if(($current_guild[guild] == None) OR ($current_guild[guild] != None))	 {
+		
+		mysql_query("INSERT INTO empnews (date, news, yourid) VALUES	('$clock', '<font class=blue>You were rejected from $guild_info[gname].</font>', '$auserid') ");
+		mysql_query("DELETE FROM guildrequests WHERE applicant='$auserid' AND gl_userid='$userid'");
 
-	echo "<div align=center><font class=yellow><b>$applicant_ename[ename] has been rejected from the guild!</b></font></div>";
-	die();
+		echo "<div align=center><font class=yellow><b>$applicant_ename[ename] has been rejected from the guild!</b></font></div>";
+		die();
+	}
 }
 
 ################
@@ -100,17 +110,20 @@ else	{
 ################
 ##	kick member
 ################
+
 if(!IsSet($kick))	{
 	include("include/S_GKICK.php");
 }
-else	{	
-
+else	{
 	// retrieve necessary guild info
 	$guild_info_query = mysql_db_query($dbnam, "SELECT * FROM guild WHERE owner='$userid'");
 		$guild_info = mysql_fetch_array($guild_info_query);
 	// get their empire name
 	$empire_info_query = mysql_db_query($dbnam, "SELECT ename FROM user WHERE userid='$remp'");
 		$empire_info = mysql_fetch_array($empire_info_query);
+	// get their current guild
+	$current_guild_query = mysql_db_query($dbnam, "SELECT guild FROM user WHERE userid='$remp'");
+		$current_guild = mysql_fetch_array($current_guild_query);
 
 	if($remp == ns)		{
 		echo"<div align=center><font class=yellow>You didn't select anybody to remove!</font></div>";
@@ -124,21 +137,29 @@ else	{
 		include("include/S_GDEL.php");
 		die();
 	}
+	elseif($empireguild != $current_guild[guild])	{	
+		echo "<div align=center><font class=yellow>That person is curently in another guild!</font></div>";
+		include("include/S_GKICK.php");
+		include("include/S_GDEL.php");
+		die();
+	}
+	else	{
 
-	$new_mem = $guild_info[mem] - 1;
+		$new_mem = $guild_info[mem] - 1;
 	
-	mysql_query("UPDATE user SET guild='None' WHERE userid='$remp'");		// update their guild membership
-	mysql_query("UPDATE guild SET mem='$new_mem' WHERE owner='$userid'");	// update number of members
-	mysql_query("DELETE FROM barter WHERE seller='$empire_info[0]' AND guild='$empireguild'");	 //	 remove any items they placed on the guild barter
+		mysql_query("UPDATE user SET guild='None' WHERE userid='$remp'");
+		mysql_query("UPDATE guild SET mem='$new_mem' WHERE owner='$userid'");	
+		mysql_query("DELETE FROM barter WHERE seller='$empire_info[0]' AND guild='$empireguild'");
 	
-	mysql_query("INSERT INTO empnews (date, news, yourid) VALUES	('$clock', '<font class=blue>You were removed from $guild_info[gname].</font>', '$remp') ");
+		mysql_query("INSERT INTO empnews (date, news, yourid) VALUES	('$clock', '<font class=blue>You were removed from $guild_info[gname].</font>', '$remp') ");
 	
-	mysql_query("INSERT INTO guildnews (date, news, guildid) VALUES	('$clock', '<font class=blue>$empire_info[0] has been removed from the guild.</font>', '$guild_info[gid]') ");
+		mysql_query("INSERT INTO guildnews (date, news, guildid) VALUES	('$clock', '<font class=blue>$empire_info[0] has been removed from the guild.</font>', '$guild_info[gid]') ");
 
-	echo"<div align=center><font class=yellow>$empire_info[0] has been removed from your guild.</font></div>";
-	include("include/S_GKICK.php");
-	include("include/S_GDEL.php");
-	die();
+		echo"<div align=center><font class=yellow>$empire_info[0] has been removed from your guild.</font></div>";
+		include("include/S_GKICK.php");
+		include("include/S_GDEL.php");
+		die();
+	}
 }
 
 ################
@@ -152,21 +173,14 @@ else	{
 	if($C_pass == $cpw)	 {
 
 		//	determine guild name
-		$GuildName = mysql_db_query($dbnam, "SELECT gname FROM guild WHERE owner='$userid'");
-		$GN = mysql_result($GuildName,"GN");
-
+			$GuildName = mysql_db_query($dbnam, "SELECT gname FROM guild WHERE owner='$userid'");
+			$GN = mysql_result($GuildName,"GN");
 		//	determine guild id
-		$GuildID = mysql_db_query($dbnam, "SELECT gid FROM guild WHERE owner='$userid'");
-		$GIDD = mysql_result($GuildID,"GIDD");
-
-		$tblname = "$GN" . "main" ."$GIDD";
-		$tblname2 = "$GN" . "msgs" . "$GIDD";	
+			$GuildID = mysql_db_query($dbnam, "SELECT gid FROM guild WHERE owner='$userid'");
+			$GIDD = mysql_result($GuildID,"GIDD");
 
 		mysql_query("UPDATE user SET guild='None' WHERE guild='$GN'");
 		mysql_query("DELETE FROM guild WHERE owner='$userid'");
-		mysql_query("DROP TABLE $tblname");
-		mysql_query("DROP TABLE $tblname2");
-
 			
 		echo "<div align=center><font class=yellow>Your guild has been deleted.</font></div>";
 		include("include/S_GDEL.php");
